requirements.txt
twitchio
python-dotenv
gTTS
python-vlc
pyyaml